Setup 3 tables in database corresponding to the classes
Unfortunately, I couldn’t get my android phone working with the
compiler. I think the usb cable I have is damaged.  Will be going to get
one today. Next I want to build a insert_PlantNode function which will
populate the a line in the table PlantNodes in database. Also build the
inverse of that, pulling a plantNode from Database to Andriod.